Initial commit
Did quite a bit of work offline. The first online commit. Have basic
fusion function working, but need to complete all the demon and fusion
rule files.

Need to work in special fusion, and add elements to the basic fusion
function

After that, need to find a way to search for chains of fusions, given a
specific skill set for a desired demon