Completed element down-fusion and up-fusion
Normal fusions now use elements as well, during fusion pair computation

Now, basic fusion is complete. All that’s left as far as having it
functionally complete is completing all the files. The files in need of
completion are:

-Demon database files. Add all races that have yet to be completed
-Fusion rules/combinations file
-Special fusion file

Other than that, an interface should be provided for specifying a
fusion to be performed, for an end-user of the program

Next step (and the real work) is to complete the fusion chain
calculation.