More changes; logical error noticed
Added interface to runner class, so fusions can be computed by user
(need to make a GUI)

The compendium fusion is not going to work 100% correctly. There is a
lot more work to be done there, and a lot more to be considered when
using the compendium. For now, the program works fine using base
demons, but the compendium functionality will need to be corrected and
improved, which will entail a lot of expansion work

The main two things left are:
1. Designing a GUI
2. Implementing working compendium functionality (lots of work here)

Also, the special fusions file needs to be completed

Additionally, I should return more than one fusion chain (ideally all,
but that could cause a runtime explosion)