Added support for compendium demons
Fusion chain calculator now uses compendium demons successfully. It
will use a compendium demon if the compendium counterpart of the demon
provides more skills (that are desired in the end result demon) than
the base counterpart.

The use of the compendium can (as always) be switched on and off at the
user’s will