Added changeable skill threshold
Added a skill threshold for evaluating whether or not to use compendium
demons. So a compendium demon will only be used if it provides more
skills desired in the final demon than its base counterpart, AND if the
number of skills it provides matches or surpasses a skill threshold.

The skill threshold can be changed at the user’s discretion