Created new directory for Apocalypse
Data files can now be entered into the relevant sections for
Apocalypse, and the code should all work just the same.

The only code that needs to change is the Race enum, which needs to be
updated with a new list to match Apocalypse races. Other than that,
it’s all the same